screenshots/requirements.txt
Appium-Python-Client==3.1.0
pytest==8.1.1
pytest-html==4.1.1
Pillow==10.3.0
python-dotenv==1.0.1
adbutils==2.6.0

